fix: arm64 build and old kernel versions

+++ driver/modern_bpf/programs/tail_called/events/syscall_dispatched_events/clone.bpf.c
@@ -236,7 +236,6 @@ int BPF_PROG(t2_clone_x,
 	     struct pt_regs *regs,
 	     long ret)
 {
-
 	struct auxiliary_map *auxmap = auxmap__get();
 	if(!auxmap)
 	{
@@ -248,7 +247,7 @@ int BPF_PROG(t2_clone_x,
 	struct task_struct *task = get_current_task();
 
 	/* Parameter 21: pid_namespace init task start_time monotonic time in ns (type: PT_UINT64) */
-	auxmap__store_u64_param(auxmap, (u64)extract__task_pidns_start_time(task, PIDTYPE_TGID, ret));
+	auxmap__store_u64_param(auxmap, extract__task_pidns_start_time(task, PIDTYPE_TGID, ret));
 
 	/*=============================== COLLECT PARAMETERS  ===========================*/
